32 bit sparse lists

+++ tmu/lib/include/ClauseBankSparse.h
@@ -44,8 +44,8 @@ void cbs_calculate_clause_outputs_predict_packed_X(
     int number_of_clauses,
     int number_of_literals,
     unsigned int *clause_output_batch,
-    unsigned short *clause_bank_included,
-    unsigned short *clause_bank_included_length
+    unsigned int *clause_bank_included,
+    unsigned int *clause_bank_included_length
 );
 
 void cbs_unpack_clause_output(
@@ -70,8 +70,8 @@ void cbs_calculate_clause_outputs_update(
     int number_of_clauses,
     int number_of_literals,
     unsigned int *clause_output,
-    unsigned short *clause_bank_included,
-    unsigned short *clause_bank_included_length
+    unsigned int *clause_bank_included,
+    unsigned int *clause_bank_included_length
 );
 
 void cbs_calculate_clause_outputs_predict(
@@ -79,8 +79,8 @@ void cbs_calculate_clause_outputs_predict(
     int number_of_clauses,
     int number_of_literals,
     unsigned int *clause_output,
-    unsigned short *clause_bank_included,
-    unsigned short *clause_bank_included_length
+    unsigned int *clause_bank_included,
+    unsigned int *clause_bank_included_length
 );
 
 void cbs_type_i_feedback(
@@ -97,12 +97,12 @@ void cbs_type_i_feedback(
     int number_of_clauses,
     int number_of_literals,
     int number_of_states,
-    unsigned short *clause_bank_included,
-    unsigned short *clause_bank_included_length,
-    unsigned short *clause_bank_excluded,
-    unsigned short *clause_bank_excluded_length,
-    unsigned short *clause_bank_unallocated,
-    unsigned short *clause_bank_unallocated_length
+    unsigned int *clause_bank_included,
+    unsigned int *clause_bank_included_length,
+    unsigned int *clause_bank_excluded,
+    unsigned int *clause_bank_excluded_length,
+    unsigned int *clause_bank_unallocated,
+    unsigned int *clause_bank_unallocated_length
 );
 
 void cbs_type_ii_feedback(
@@ -114,8 +114,8 @@ void cbs_type_ii_feedback(
     int number_of_clauses,
     int number_of_literals,
     int number_of_states,
-    unsigned short *clause_bank_included,
-    unsigned short *clause_bank_included_length,
-    unsigned short *clause_bank_excluded,
-    unsigned short *clause_bank_excluded_length
+    unsigned int *clause_bank_included,
+    unsigned int *clause_bank_included_length,
+    unsigned int *clause_bank_excluded,
+    unsigned int *clause_bank_excluded_length
 );
